requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
ultralytics
pyyaml
opencv-python==4.8.0.76
torch==2.1.2+cu121
torchvision==0.16.2+cu121
numpy<2
albumentations
pandas
scikit-learn
ensemble-boxes
sahi